requirement.txt
Django==2.2
django-filter==2.1.0
django-split-settings==0.3.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Markdown==3.1
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
